REF: SelectFee
Use builtin detents in RNav 7 isntead of using a dep package
